+++ 42/media/lua/shared/PhobosLib.lua
@@ -5,7 +5,7 @@
 ---------------------------------------------------------------
 
 PhobosLib = PhobosLib or {}
-PhobosLib.VERSION = "1.7.0"
+PhobosLib.VERSION = "1.8.0"
 
 require "PhobosLib_Util"
 require "PhobosLib_Fluid"
@@ -17,3 +17,4 @@ require "PhobosLib_Skill"
 require "PhobosLib_Reset"
 require "PhobosLib_Validate"
 require "PhobosLib_Trading"
+require "PhobosLib_Migrate"